Change logic for starting the encoder thread
Added in_use state for exclusivity, but this needs to be improved to adhere
to v4l2 API specs, so we can have multiple openers of the device, but only
one reader (and lock out changes when they reader is active).

Moved thread start to buf prepare to make that change later easier, and also
so we can implement enc commands.

Signed-off-by: Ben Collins <bcollins@bluecherry.net>