Rework of how g723 registers with alsa
- One snd card and one device per solo card
- X number of devices per solo card based on nr_chans
- Enable/disable g723 irqs only when we have capture

We're still not working yet. Have to implement the buffer copies on
interrupt and update pointers so userspace can work.

Also, need to implement playback via the audio out port on the card,
and perhaps expose the g723 decoder? Probably need to do the latter to
implement the former.

Still not sure if g723 decode will be done in the driver so we expose
S16_LE to userspace, or if we will pump g723 directly to userspace and
leave it up to the application to decode g723. Alsa devs would probably
prefer that g723 decode be done in the application and let hardware
do it's own thing.

Signed-off-by: Ben Collins <bcollins@bluecherry.net>