Bug #878: Align MPEG/JPEG DMA reads to 4096 bytes to prevent false timeouts
We're getting DMA timeouts when reading a frame with a size not aligned to
4096 bytes, which is then mishandled and results in corrupt output, including
misordered frames, jumpy motion, and artifacting. Aligning these sizes in the
common case prevents most of those errors.

Further work is needed on other cases that may cause timeouts (such as buffer
wrap), and finding why the handling of these errors can corrupt the output
so badly.