encoder: Disable encoder IRQ during ring handler
There's no need to signal a new interrupt while we are actually
processing the ring, so disable them during the handler and
re-enable it afterwards. Also, recheck CURQ from hardware on every
loop in order to process all that we can during the handler (e.g. in
case a few more queues are processed while we are handling prior ones).

Signed-off-by: Ben Collins <bcollins@bluecherry.net>